requirements.txt
Django==2.2.4
django-appconf==1.0.3
django-select2==7.1.1
pyserial==3.4
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
